Dynamically add timeouts to chart generation
So far, we only considered timeouts of excluded schemas. Now, included schemas that run into timeout are also plotet on the charts.

+++ artifacts/charts/clean_data.r
@@ -8,18 +8,25 @@ gen_df <- function(in_path, out_path, write_timeouts=FALSE) {
     write.csv(clean_df,paste(out_path,"clean_data.csv",sep="/"), row.names = FALSE)
     if (write_timeouts) {
       # get all rows with "TimeoutException" in any column except the first
-      timeout_df <- df[apply(df[,-1], 1, function(x) any(grepl("TimeoutException", x))),]
-      # write to timeout.csv in out_path directory
-      write.csv(timeout_df,paste(out_path,"timeout.csv",sep="/"), row.names = FALSE)
+      header <- c("inSize", "objectId")
+      timeout_df <- df[apply(df[,-1], 1, function(x) any(grepl("TimeoutException", x))), header]
+      # check if timeout.csv already exists
+      if (file.exists(paste(out_path, "timeout_excluded.csv", sep="/"))) {
+        existing_timeout_df <- read.csv(paste(out_path, "timeout_excluded.csv", sep="/"))
+        new_timeout_df <- rbind(existing_timeout_df, timeout_df)
+        write.csv(new_timeout_df, paste(out_path, "timeout.csv", sep="/"), row.names = FALSE)
+      } else {
+        write.csv(timeout_df, paste(out_path, "timeout.csv", sep="/"), row.names = FALSE)
+      }
     }
   } else {
     print(paste("File not found:", in_path))
   }
 }
 
-gen_df('data/github/results.csv', 'data/github/')
-gen_df('data/kubernetes/results.csv', 'data/kubernetes/')
-gen_df('data/snowplow/results.csv', 'data/snowplow/')
-gen_df('data/wp/results.csv', 'data/wp/')
-gen_df('data/allOf_containment/results.csv', 'data/allOf_containment/')
+gen_df('data/github/results.csv', 'data/github/', write_timeouts=TRUE)
+gen_df('data/kubernetes/results.csv', 'data/kubernetes/', write_timeouts=TRUE)
+gen_df('data/snowplow/results.csv', 'data/snowplow/', write_timeouts=TRUE)
+gen_df('data/wp/results.csv', 'data/wp/', write_timeouts=TRUE)
+gen_df('data/allOf_containment/results.csv', 'data/allOf_containment/', write_timeouts=TRUE)
 gen_df('data/schemastore_containment/results.csv', 'data/schemastore_containment/', write_timeouts=TRUE)
